-Added UserPersistence class and GetUserByLoginToken method -Added Tests for GetUserByLoginToken -Added Tests for RemoveUserFromTeachingClass -Fixed initalize new Subjects, AssignedAccounts and OpenRequests list in constructor of a teaching class

+++ src/SuperSmart/SuperSmart.Core/Data/ViewModels/RemoveUserFromTeachingClassViewModel.cs
@@ -6,9 +6,9 @@ namespace SuperSmart.Core.Data.ViewModels
     public class RemoveUserFromTeachingClassViewModel
     {
         [Required(ErrorMessage = "Id for teaching class required")]
-        public int TeachingClassId { get; set; }
+        public long TeachingClassId { get; set; }
 
         [Required(ErrorMessage = "Id for user required")]
-        public int UserId { get; set; }
+        public long UserId { get; set; }
     }
 }
